+++ views/write.ejs
@@ -6,7 +6,7 @@
 
     <div class="container mt-4">
       <h2 class="banner ml-2 my-3 text-center">새로운 글</h2>
-      <form action="/add" method="POST">
+      <form action="/write/add" method="POST">
         <div class="list-group-item">
           <div class="form-group">
             <label class="title">제목</label>
